+++ patches/src/main/kotlin/app/morphe/patches/reddit/layout/sidebar/Fingerprints.kt
@@ -31,6 +31,7 @@ internal val communityDrawerPresenterFingerprint = Fingerprint(
         Opcode.MOVE_RESULT_OBJECT,
     ),
     custom = { methodDef, _ ->
+        // TODO: Convert this to an instruction filter
         indexOfKotlinCollectionInstruction(methodDef) >= 0
     }
 )
@@ -62,23 +63,19 @@ internal val redditProLoaderFingerprint = Fingerprint(
 )
 
 internal val sidebarComponentsPatchFingerprint = Fingerprint(
+    definingClass = EXTENSION_CLASS_DESCRIPTOR,
+    name = "getHeaderItemName",
     accessFlags = listOf(AccessFlags.PRIVATE, AccessFlags.STATIC),
-    returnType = "Ljava/lang/String;",
-    custom = { methodDef, classDef ->
-        methodDef.name == "getHeaderItemName" &&
-                classDef.type.endsWith("/SidebarComponentsPatch;")
-    }
+    returnType = "Ljava/lang/String;"
 )
 
 internal val headerItemUiModelToStringFingerprint = Fingerprint(
+    name = "toString",
     returnType = "Ljava/lang/String;",
     strings = listOf(
         "HeaderItemUiModel(uniqueId=",
         ", type="
-    ),
-    custom = { methodDef, _ ->
-        methodDef.name == "toString"
-    }
+    )
 )
 
 // TODO: Replace with fieldAccess() instruction filter usage.