+++ patches/src/main/kotlin/app/morphe/patches/reddit/layout/sidebar/SidebarComponentsPatch.kt
@@ -23,7 +23,7 @@ import com.android.tools.smali.dexlib2.iface.instruction.ReferenceInstruction
 import com.android.tools.smali.dexlib2.iface.reference.FieldReference
 import com.android.tools.smali.dexlib2.immutable.ImmutableMethod
 
-private const val EXTENSION_CLASS_DESCRIPTOR =
+internal const val EXTENSION_CLASS_DESCRIPTOR =
     "Lapp/morphe/extension/reddit/patches/SidebarComponentsPatch;"
 
 @Suppress("unused")
@@ -121,7 +121,7 @@ val sidebarComponentsPatch = bytecodePatch(
         } else {
             val communityDrawerPresenterMethod =
                 communityDrawerPresenterFingerprint.match(
-                    mutableClassDefBy(communityDrawerPresenterConstructorFingerprint.classDef)
+                    communityDrawerPresenterConstructorFingerprint.classDef
                 ).method
 
             fun getDrawerField(