+++ patches/src/main/kotlin/app/morphe/patches/reddit/misc/extension/hooks/ApplicationInitHook.kt
@@ -3,15 +3,13 @@ package app.morphe.patches.reddit.misc.extension.hooks
 import app.morphe.patcher.Fingerprint
 
 internal val redditMainActivityOnCreateFingerprint = Fingerprint(
+    definingClass = "Lcom/reddit/launch/main/MainActivity;",
+    name = "onCreate",
     returnType = "V",
-    parameters = listOf("Landroid/os/Bundle;"),
-    custom = { method, classDef ->
-        method.name == "onCreate" && classDef.type == "Lcom/reddit/launch/main/MainActivity;"
-    }
+    parameters = listOf("Landroid/os/Bundle;")
 )
 
 internal val redditActivityOnCreateFingerprint = Fingerprint(
-    custom = { method, classDef ->
-        method.name == "onCreate" && classDef.type.endsWith("/FrontpageApplication;")
-    }
+    definingClass = "/FrontpageApplication;",
+    name = "onCreate"
 )